+++ src/main/java/com/zach2039/factorymade/data/FactoryMadeRecipeProvider.java
@@ -44,20 +44,22 @@ public class FactoryMadeRecipeProvider extends RecipeProvider {
 		
 		// Iron Non-Slip Walkway
 		{
-			IndustrialShaperRecipeBuilder.industrialshaper(Ingredient.of(Blocks.IRON_BLOCK), ModBlocks.IRON_NON_SLIP_WALKWAY.get(), 16)
-						.unlockedBy("has_iron_block", has(Blocks.IRON_BLOCK))
-						.save(recipeConsumer, new ResourceLocation(FactoryMade.MODID, "iron_non_slip_walkway"));			
-			addSlabRecipes(recipeConsumer, ModBlocks.IRON_NON_SLIP_WALKWAY.get(), ModBlocks.IRON_NON_SLIP_WALKWAY_SLAB.get());
-			addStairRecipes(recipeConsumer, ModBlocks.IRON_NON_SLIP_WALKWAY.get(), ModBlocks.IRON_NON_SLIP_WALKWAY_STAIRS.get());
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.IRON_BLOCK, 
+					ModBlocks.IRON_NON_SLIP_WALKWAY.get(), 16, 
+					ModBlocks.IRON_NON_SLIP_WALKWAY_STAIRS.get(),
+					ModBlocks.IRON_NON_SLIP_WALKWAY_SLAB.get()
+					);
 		}
 		
 		// Rusted Iron Non-Slip Walkway
 		{
-			IndustrialShaperRecipeBuilder.industrialshaper(Ingredient.of(Blocks.IRON_BLOCK), ModBlocks.RUSTED_IRON_NON_SLIP_WALKWAY.get(), 16)
-						.unlockedBy("has_iron_block", has(Blocks.IRON_BLOCK))
-						.save(recipeConsumer, new ResourceLocation(FactoryMade.MODID, "rusted_iron_non_slip_walkway"));			
-			addSlabRecipes(recipeConsumer, ModBlocks.RUSTED_IRON_NON_SLIP_WALKWAY.get(), ModBlocks.RUSTED_IRON_NON_SLIP_WALKWAY_SLAB.get());
-			addStairRecipes(recipeConsumer, ModBlocks.RUSTED_IRON_NON_SLIP_WALKWAY.get(), ModBlocks.RUSTED_IRON_NON_SLIP_WALKWAY_STAIRS.get());
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.IRON_BLOCK, 
+					ModBlocks.RUSTED_IRON_NON_SLIP_WALKWAY.get(), 16, 
+					ModBlocks.RUSTED_IRON_NON_SLIP_WALKWAY_STAIRS.get(),
+					ModBlocks.RUSTED_IRON_NON_SLIP_WALKWAY_SLAB.get()
+					);
 		}
 		
 		// Iron Truss
@@ -74,22 +76,91 @@ public class FactoryMadeRecipeProvider extends RecipeProvider {
 						.save(recipeConsumer, new ResourceLocation(FactoryMade.MODID, "rusted_iron_truss"));			
 		}
 		
-		// White Cinder Block Bricks
+		// Cinder Block Bricks
 		{
-			IndustrialShaperRecipeBuilder.industrialshaper(Ingredient.of(Blocks.WHITE_CONCRETE), ModBlocks.WHITE_CINDER_BLOCK_BRICKS.get(), 16)
-						.unlockedBy("has_white_concrete_block", has(Blocks.WHITE_CONCRETE))
-						.save(recipeConsumer, new ResourceLocation(FactoryMade.MODID, "white_cinder_block_bricks"));			
-			addSlabRecipes(recipeConsumer, ModBlocks.WHITE_CINDER_BLOCK_BRICKS.get(), ModBlocks.WHITE_CINDER_BLOCK_BRICKS_SLAB.get());
-			addStairRecipes(recipeConsumer, ModBlocks.WHITE_CINDER_BLOCK_BRICKS.get(), ModBlocks.WHITE_CINDER_BLOCK_BRICKS_STAIRS.get());
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.WHITE_CONCRETE, 
+					ModBlocks.WHITE_CINDER_BLOCK_BRICKS.get(), 4, 
+					ModBlocks.WHITE_CINDER_BLOCK_BRICKS_STAIRS.get(),
+					ModBlocks.WHITE_CINDER_BLOCK_BRICKS_SLAB.get()
+					);
+			
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.YELLOW_CONCRETE, 
+					ModBlocks.YELLOW_CINDER_BLOCK_BRICKS.get(), 4, 
+					ModBlocks.YELLOW_CINDER_BLOCK_BRICKS_STAIRS.get(),
+					ModBlocks.YELLOW_CINDER_BLOCK_BRICKS_SLAB.get()
+					);
+			
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.RED_CONCRETE, 
+					ModBlocks.RED_CINDER_BLOCK_BRICKS.get(), 4, 
+					ModBlocks.RED_CINDER_BLOCK_BRICKS_STAIRS.get(),
+					ModBlocks.RED_CINDER_BLOCK_BRICKS_SLAB.get()
+					);
+			
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.BLUE_CONCRETE, 
+					ModBlocks.BLUE_CINDER_BLOCK_BRICKS.get(), 4, 
+					ModBlocks.BLUE_CINDER_BLOCK_BRICKS_STAIRS.get(),
+					ModBlocks.BLUE_CINDER_BLOCK_BRICKS_SLAB.get()
+					);
+			
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.GREEN_CONCRETE, 
+					ModBlocks.GREEN_CINDER_BLOCK_BRICKS.get(), 4, 
+					ModBlocks.GREEN_CINDER_BLOCK_BRICKS_STAIRS.get(),
+					ModBlocks.GREEN_CINDER_BLOCK_BRICKS_SLAB.get()
+					);
+			
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.BROWN_CONCRETE, 
+					ModBlocks.BROWN_CINDER_BLOCK_BRICKS.get(), 4, 
+					ModBlocks.BROWN_CINDER_BLOCK_BRICKS_STAIRS.get(),
+					ModBlocks.BROWN_CINDER_BLOCK_BRICKS_SLAB.get()
+					);
 		}
 		
-		// Black Asbestos Tiles
+		// Concrete Siding
 		{
-			IndustrialShaperRecipeBuilder.industrialshaper(Ingredient.of(Blocks.BLACK_CONCRETE), ModBlocks.BLACK_ASBESTOS_TILES.get(), 16)
-						.unlockedBy("has_black_concrete_block", has(Blocks.BLACK_CONCRETE))
-						.save(recipeConsumer, new ResourceLocation(FactoryMade.MODID, "black_asbestos_tiles"));			
-			addSlabRecipes(recipeConsumer, ModBlocks.BLACK_ASBESTOS_TILES.get(), ModBlocks.BLACK_ASBESTOS_TILES_SLAB.get());
-			addStairRecipes(recipeConsumer, ModBlocks.BLACK_ASBESTOS_TILES.get(), ModBlocks.BLACK_ASBESTOS_TILES_STAIRS.get());
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.WHITE_CONCRETE, 
+					ModBlocks.CONCRETE_SIDING.get(), 4, 
+					ModBlocks.CONCRETE_SIDING_STAIRS.get(),
+					ModBlocks.CONCRETE_SIDING_SLAB.get()
+					);
+			
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.WHITE_CONCRETE, 
+					ModBlocks.WEATHERED_CONCRETE_SIDING.get(), 4, 
+					ModBlocks.WEATHERED_CONCRETE_SIDING_STAIRS.get(),
+					ModBlocks.WEATHERED_CONCRETE_SIDING_SLAB.get()
+					);
+		}
+		
+		// Asbestos Tiles
+		{
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.BLACK_CONCRETE, 
+					ModBlocks.BLACK_ASBESTOS_TILES.get(), 4, 
+					ModBlocks.BLACK_ASBESTOS_TILES_STAIRS.get(),
+					ModBlocks.BLACK_ASBESTOS_TILES_SLAB.get()
+					);
+			
+			addBlockStairSlab(recipeConsumer, 
+					Blocks.WHITE_CONCRETE, 
+					ModBlocks.WHITE_ASBESTOS_TILES.get(), 4, 
+					ModBlocks.WHITE_ASBESTOS_TILES_STAIRS.get(),
+					ModBlocks.WHITE_ASBESTOS_TILES_SLAB.get()
+					);
+		}
+		
+		// Fluorescent Light Panel
+		{
+			IndustrialShaperRecipeBuilder.industrialshaper(Ingredient.of(Blocks.GLOWSTONE), ModBlocks.FLUORESCENT_LIGHT_PANEL.get(), 2)
+						.unlockedBy("has_glowstone", has(Blocks.GLOWSTONE))
+						.save(recipeConsumer, new ResourceLocation(FactoryMade.MODID, "fluorescent_light_panel"));			
+			addSlabRecipes(recipeConsumer, ModBlocks.FLUORESCENT_LIGHT_PANEL.get(), ModBlocks.FLUORESCENT_LIGHT_PANEL_SLAB.get());
 		}
 	}
 
@@ -117,6 +188,14 @@ public class FactoryMadeRecipeProvider extends RecipeProvider {
 		IndustrialShaperRecipeBuilder.industrialshaper(Ingredient.of(inputBlock), outputStair, 1)
 						.unlockedBy("has_block", has(inputBlock))
 						.save(recipeConsumer, new ResourceLocation(FactoryMade.MODID, ModRegistryUtil.getKey(inputBlock).getPath() + "_stair_from_shaper"));
+	}
+	
+	private void addBlockStairSlab(final Consumer<FinishedRecipe> recipeConsumer, Block inputBlock, Block outputBlock, int count, StairBlock outputStair, SlabBlock outputSlab) {
+		IndustrialShaperRecipeBuilder.industrialshaper(Ingredient.of(inputBlock), outputBlock, count)
+				.unlockedBy("has_block", has(inputBlock))
+				.save(recipeConsumer, new ResourceLocation(FactoryMade.MODID, ModRegistryUtil.getKey(outputBlock).getPath()));			
+		addStairRecipes(recipeConsumer, outputBlock, outputStair);
+		addSlabRecipes(recipeConsumer, outputBlock, outputSlab);
 	}
 	
 	@Override