+++ src/main/java/com/zach2039/factorymade/data/loot/FactoryMadeBlockLootTables.java
@@ -31,9 +31,44 @@ public class FactoryMadeBlockLootTables extends BlockLoot {
 		dropSelf(ModBlocks.WHITE_CINDER_BLOCK_BRICKS_STAIRS.get());
 		dropSelf(ModBlocks.WHITE_CINDER_BLOCK_BRICKS_SLAB.get());
 		
+		dropSelf(ModBlocks.YELLOW_CINDER_BLOCK_BRICKS.get());
+		dropSelf(ModBlocks.YELLOW_CINDER_BLOCK_BRICKS_STAIRS.get());
+		dropSelf(ModBlocks.YELLOW_CINDER_BLOCK_BRICKS_SLAB.get());
+		
+		dropSelf(ModBlocks.RED_CINDER_BLOCK_BRICKS.get());
+		dropSelf(ModBlocks.RED_CINDER_BLOCK_BRICKS_STAIRS.get());
+		dropSelf(ModBlocks.RED_CINDER_BLOCK_BRICKS_SLAB.get());
+		
+		dropSelf(ModBlocks.BLUE_CINDER_BLOCK_BRICKS.get());
+		dropSelf(ModBlocks.BLUE_CINDER_BLOCK_BRICKS_STAIRS.get());
+		dropSelf(ModBlocks.BLUE_CINDER_BLOCK_BRICKS_SLAB.get());
+		
+		dropSelf(ModBlocks.GREEN_CINDER_BLOCK_BRICKS.get());
+		dropSelf(ModBlocks.GREEN_CINDER_BLOCK_BRICKS_STAIRS.get());
+		dropSelf(ModBlocks.GREEN_CINDER_BLOCK_BRICKS_SLAB.get());
+		
+		dropSelf(ModBlocks.BROWN_CINDER_BLOCK_BRICKS.get());
+		dropSelf(ModBlocks.BROWN_CINDER_BLOCK_BRICKS_STAIRS.get());
+		dropSelf(ModBlocks.BROWN_CINDER_BLOCK_BRICKS_SLAB.get());
+		
+		dropSelf(ModBlocks.CONCRETE_SIDING.get());
+		dropSelf(ModBlocks.CONCRETE_SIDING_STAIRS.get());
+		dropSelf(ModBlocks.CONCRETE_SIDING_SLAB.get());
+		
+		dropSelf(ModBlocks.WEATHERED_CONCRETE_SIDING.get());
+		dropSelf(ModBlocks.WEATHERED_CONCRETE_SIDING_STAIRS.get());
+		dropSelf(ModBlocks.WEATHERED_CONCRETE_SIDING_SLAB.get());
+		
 		dropSelf(ModBlocks.BLACK_ASBESTOS_TILES.get());
 		dropSelf(ModBlocks.BLACK_ASBESTOS_TILES_STAIRS.get());
 		dropSelf(ModBlocks.BLACK_ASBESTOS_TILES_SLAB.get());
+		
+		dropSelf(ModBlocks.WHITE_ASBESTOS_TILES.get());
+		dropSelf(ModBlocks.WHITE_ASBESTOS_TILES_STAIRS.get());
+		dropSelf(ModBlocks.WHITE_ASBESTOS_TILES_SLAB.get());
+		
+		dropSelf(ModBlocks.FLUORESCENT_LIGHT_PANEL.get());
+		dropSelf(ModBlocks.FLUORESCENT_LIGHT_PANEL_SLAB.get());
 	}
 
 	@Override