+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+### v1.19-1.0.2-3
+
+ - Fix JEI category using Stonecutter as icon 
+ - Add white asbestos tiles
+ - De-grunge black asbestos tiles texture slightly
+ - Add fluorescent light panels
+ - Add more cinder block brick colors
+ - Add weathered and unweathered concrete siding
+
+### v1.19-1.0.1-2
+
+ - Reduce bricks and tiles recipe output to 4 from 16
+ - Change dependency to JEI from REI in mods.toml
+
 ### v1.19-1.0.0-1
 
  - Initial release